# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tracklet API Documentation",
        "description": "Multi-Organization Management System API - Complete API documentation for mobile and web developers",
        "contact": {
            "email": "support@tracklet.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "API Server (uses current host)"
        }
    ],
    "paths": {
...
        "/api/inventory/items": {
            "get": {
                "tags": [
                    "Inventory"
                ],
                "summary": "List inventory items",
                "operationId": "c94299512b4543730b135ec5e6a7e34f",
                "parameters": [
                    {
                        "name": "category",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "low_stock",
                        "in": "query",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of inventory items"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Inventory"
                ],
                "summary": "Create inventory item",
                "operationId": "62d2d703d27eb48f941dfed9d56d5707",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "quantity",
                                    "minimum_threshold",
                                    "unit_price"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "A4 Paper"
                                    },
                                    "category": {
                                        "type": "string",
                                        "example": "Office Supplies"
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "example": 100
                                    },
                                    "minimum_threshold": {
                                        "type": "integer",
                                        "example": 20
                                    },
                                    "unit_price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": "5.00"
                                    },
                                    "unit": {
                                        "type": "string",
                                        "example": "reams"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Item created successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/inventory/items/{id}": {
            "get": {
                "tags": [
                    "Inventory"
                ],
                "summary": "Get inventory item",
                "operationId": "890207c5894190058e116500661b7db3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Item details"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Inventory"
                ],
                "summary": "Update inventory item",
                "operationId": "375f6381804ca69f92ed79947b430cbb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "quantity",
                                    "minimum_threshold",
                                    "unit_price"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "category": {
                                        "type": "string"
                                    },
                                    "quantity": {
                                        "type": "integer"
                                    },
                                    "minimum_threshold": {
                                        "type": "integer"
                                    },
                                    "unit_price": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "unit": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Item updated successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Inventory"
                ],
                "summary": "Delete inventory item",
                "operationId": "80c5bafcc7914dbc8ed1f150a2bd2821",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Item deleted successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/inventory/items/{id}/transactions": {
            "get": {
                "tags": [
                    "Inventory"
                ],
                "summary": "Get stock transactions",
                "operationId": "379db8f71194df757d66ce8f52ccb042",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "in",
                                "out"
                            ]
                        }
                    },
                    {
                        "name": "date_from",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "date_to",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Stock transactions"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/inventory/low-stock": {
            "get": {
                "tags": [
                    "Inventory"
                ],
                "summary": "Get low stock items",
                "operationId": "2cecd4276ee4df060a9d69cc0a72828a",
                "responses": {
                    "200": {
                        "description": "Low stock items"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/inventory/aging-report": {
            "get": {
                "tags": [
                    "Inventory"
                ],
                "summary": "Get item aging report",
                "operationId": "a710d114390a260a9b1979bd7a5a6378",
                "responses": {
                    "200": {
                        "description": "Aging report"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}